HTTP status codes on game restore trap error spawning frotz and report with status code detect invalid game pid on game delete and send 404

+++ src/server.js
@@ -35,6 +35,11 @@ app.post('/games', function(req, res) {
     var game = req.body.game.replace(/[^a-z0-9]/, '');
     var zFile = __dirname + '/../zcode/' + game + '.z5';
     var child = spawn(__dirname + "/../frotz/dfrotz", ["-S 0", zFile]);
+    child.on('error', function(err) {
+      console.log('Error spawning game: ' + err);
+      res.status(500).send({error: 'Could not create the game.'});
+      return;
+    });
 
     console.log("Game %s spawned for %s: %s", child.pid, label, zFile)
     games[child.pid] = {
@@ -70,9 +75,13 @@ app.get('/games', function(req, res) {
 
 app.delete('/games/:pid', function(req, res) {
     var pid = req.params.pid;
-    games[pid].process.kill();
-    delete games[pid];
-    res.send("Game for " + pid + " terminated.");
+    if (undefined === games[pid]) {
+      res.status(404).send({error: 'Game ' + pid + ' not found.'});
+    } else {
+      games[pid].process.kill();
+      delete games[pid];
+      res.send("Game for " + pid + " terminated.");
+    }
 });
 
 app.post('/games/:pid/action', function(req, res) {
@@ -114,6 +123,7 @@ app.post('/games/:pid/save', function(req, res) {
             if (err) {
                 console.log("S3 Error: %j", err);
                 data = " (not saved to cloud) " + err;
+                res.status(503);
             } else {
                 console.log("S3 success: %j", updata);
                 data = data + " (saved to cloud)";
@@ -192,12 +202,12 @@ app.post('/games/:pid/restore', function(req, res) {
             pid: pid,
             data: 'Failed to find save game' + data
         }
-        res.send(response)
+        res.status(404).send(response)
     };
 
     var getFromS3 = function() {
         if (s3 === undefined) {
-            res.send('Failed to find game on Disk\nCannot get from S3: not configured');
+            res.status(404).send('Failed to find game on Disk\nCannot get from S3: not configured');
             return;
         }
         // Grab from S3
